对vcb-studio进行更进一步的修复; parser info里episode改成float; 添加对jsum的简单的匹配（待测试）

+++ Banned.AniParser/Models/ParserInfo.cs
@@ -17,7 +17,7 @@ public class ParserInfo
     /// <summary>
     /// 单集集数
     /// </summary>
-    public int? Episode { get; init; }
+    public float? Episode { get; init; }
 
     /// <summary>
     /// 多集的第一集
